+++ Changelog.md
@@ -1,3 +1,6 @@
+# Master
+  * Generate unique ifplId if it's missing from the XML file [#109](devteamreims/4ME#109)
+
 # v0.1.3
   * Add `npm run build:docker` command to build a properly tagged docker image
   * **BREAKING**: Remove `NODE_TLS_REJECT_UNAUTHORIZED` from the source code. To allow non valid SSL certs, please set this variable to `0` in your env